// src/app/models/objective.model.ts
// filepath: c:\Users\INMSA ARGO\Proyectos\qsamaxfro\qsamaxfro\qsafrotnend\src\app\models\objective.model.ts
export interface Objective {
    id: number;
    name: string;
    description: string;
    goalBase: string;
    goalAverage: string;
    active: boolean;
    departmentId: number;
    usercreate: string;
    createdate: string;
}